requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
motor==3.4.0
redis==5.0.4
python-dotenv==1.0.1
pydantic==2.7.1
pymongo==4.7.2